Small fixes (report typo, audio file name added only when played)

+++ Source/GUI/GUIStats.h
@@ -302,6 +302,6 @@ private:
     std::vector<Label*> mVecFreqLabels;
     std::vector<Label*> mVecCountLabels;
     std::vector<StatsBar*> mVecBars;
-    std::vector<float> vecStartXPos =
+    std::vector<double> vecStartXPos =
         {0.3, 0.35, 0.4, 0.45, 0.5, 0.55, 0.6, 0.65, 0.7, 0.75};
 };
